backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
python-multipart==0.0.6
gspread==6.0.2
google-auth==2.27.0
